Split up actor classes
Uses a trick from the PF2E FoundryVTT system to allow actor subclasses
in Foundry.
