+++ src/module/actor/index.js
@@ -1,4 +1,4 @@
-export { TorchbearerCharacterSheet } from "./character";
-export { TorchbearerMonsterSheet } from "./monster";
-export { TorchbearerNPCSheet } from "./npc";
 export { TorchbearerBaseActor } from "./base-actor";
+export { TorchbearerCharacterActor, TorchbearerCharacterSheet } from "./character";
+export { TorchbearerMonsterActor, TorchbearerMonsterSheet } from "./monster";
+export { TorchbearerNPCActor, TorchbearerNPCSheet } from "./npc";
