+++ src/module/actor/monster/actor.js
@@ -1,0 +1,3 @@
+import { TorchbearerBaseActor } from "../base-actor";
+
+export class TorchbearerMonsterActor extends TorchbearerBaseActor {}
