+++ src/module/actor/monster/index.js
@@ -1,1 +1,2 @@
+export { TorchbearerMonsterActor } from "./actor";
 export { TorchbearerMonsterSheet } from "./sheet";
