+++ src/module/actor/npc/actor.js
@@ -1,0 +1,3 @@
+import { TorchbearerCharacterActor } from "../character";
+
+export class TorchbearerNPCActor extends TorchbearerCharacterActor {}
