+++ src/module/actor/npc/index.js
@@ -1,1 +1,2 @@
+export { TorchbearerNPCActor } from "./actor";
 export { TorchbearerNPCSheet } from "./sheet";
